requirements.txt
streamlit==0.89.0
plotly==5.3.1
pandas==1.3.3
numpy==1.21.5
pillow==8.3.2
streamlit_autorefresh==0.1.4
streamlit_folium==0.3.0
folium==0.12.2
requests==2.26.0
matplotlib==3.4.3
windrose==1.6.9

